Moved category components to own file
