+++ src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Link } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import axios from 'axios';
 import './index.css';
+import Categories from './category.jsx';
 
 class SearchForm extends React.Component {
   constructor(props) {
@@ -52,39 +53,6 @@ class SearchResult extends React.Component {
         <p>{this.props.record.LATITUDE} {this.props.record.LONGITUDE}</p>
         <p>{this.props.record.ORGANISATION_PURPOSE}</p>
       </div>
-    );
-  }
-}
-
-class CategoryLink extends React.Component {
-  render() {
-    return (
-      <li className="category-link">
-        <Link to={this.props.record.name} onClick={this.handleCategorySelection}>
-          {this.props.record.name}
-        </Link>
-      </li>
-    );
-  }
-}
-
-class Categories extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {categories: []};
-    this.fetchCategories();
-  }
-  fetchCategories() {
-    let sql = encodeURI('SELECT DISTINCT "LEVEL_1_CATEGORY" as name FROM "35de6bf8-b254-4025-89f5-da9eb6adf9a0" ORDER BY name');
-    let url = `https://catalogue.data.govt.nz/api/3/action/datastore_search_sql?sql=${sql}`;
-    axios.get(url)
-      .then(res => {
-        this.setState({ categories: res.data.result.records });
-      });
-  }
-  render() {
-    return this.state.categories.map((record) =>
-      <CategoryLink record={record} />
     );
   }
 }
